Allow Symfony 4
Symfony 4 Beta is here. In order to be able to use Symfony 4 with API-Pack, it's required to allow Symfony v4 and also ReflectionDocBlock v4.

This is especially important because Symfony recommends API-Pack as best practice for creating APIs (pleas look at the docs for Symfony Flex and at the Symfony Recipes Server). This is why it would be a pity if API-Pack cannot be used with Symfony 4.